fix(maintainability): close terminal review findings
