+++ framework/maintainability/semantic-amplification.manifest.json
@@ -451,47 +451,71 @@
     },
     {
       "id": "semantic-authority-convergence",
-      "state": "retained-dependency",
-      "evidence": [],
+      "state": "proved",
+      "evidence": [
+        "framework/maintainability/terminal-evidence-matrix.json"
+      ],
       "dependentAssignment": {
         "assignmentId": "2026-07-26-kungfu-authority-convergence",
         "requestRoot": "sha256:3c21a18843d601a08012bce640fa6c184358179cc9f82999d7420b80d7806b97",
         "captureReceiptRoot": "sha256:e9668f6ac8cf97bb59db87c1c5ec04076dd14b4d294abc1a25408c2d2dfb552b"
       },
-      "claimBoundary": "The current report maps authority routes and detects shadow surfaces; material convergence remains a separately owned paused Assignment."
+      "closureBinding": {
+        "sealedStateRoot": "sha256:41449544ae9f64f6a6e6e7ae69f24c0fd9049e38beef4bba2bbed3e3e730e98c",
+        "matrixRow": "2026-07-26-kungfu-authority-convergence"
+      },
+      "claimBoundary": "The predecessor seal closes this Assignment; the current report remains a navigation projection over the declared authorities."
     },
     {
       "id": "responsibility-hotspot-decomposition",
-      "state": "retained-dependency",
-      "evidence": [],
+      "state": "proved",
+      "evidence": [
+        "framework/maintainability/terminal-evidence-matrix.json"
+      ],
       "dependentAssignment": {
         "assignmentId": "2026-07-26-kungfu-responsibility-hotspot-decomposition",
         "requestRoot": "sha256:5f71d58aac57e46f268dd1953473249aab182e0e3f8061a7d75929c3f81be9ee",
         "captureReceiptRoot": "sha256:9694edded6e090e588c56286ebf9c957718ecfb14b3e92b8b74e7ffbcace2bde"
       },
-      "claimBoundary": "Two bounded source-tooling seams are proved here; decomposition of the repository responsibility hotspots remains a separately owned paused Assignment."
+      "closureBinding": {
+        "sealedStateRoot": "sha256:41449544ae9f64f6a6e6e7ae69f24c0fd9049e38beef4bba2bbed3e3e730e98c",
+        "matrixRow": "2026-07-26-kungfu-responsibility-hotspot-decomposition"
+      },
+      "claimBoundary": "The predecessor seal closes the prior decomposition Assignment; newly discovered hotspots remain governed by the current exact-head terminal review."
     },
     {
       "id": "linux-native-and-product",
-      "state": "retained-dependency",
-      "evidence": [],
+      "state": "proved",
+      "evidence": [
+        "framework/maintainability/terminal-evidence-matrix.json"
+      ],
       "dependentAssignment": {
         "assignmentId": "2026-07-26-kungfu-linux-product-qualification",
         "requestRoot": "sha256:abfb20e87a7c06437431c133b4e32c5cb72a8c1ca4c7e5ddb8917b7eb77e0a4f",
         "captureReceiptRoot": "sha256:91305cbc2255ed4ef18e2f7966be6f9acbc7ea68544b1535c261a70cebfa8974"
       },
-      "claimBoundary": "Exact delivered-commit Linux native and installed-product qualification remains a separately owned paused Assignment."
+      "closureBinding": {
+        "sealedStateRoot": "sha256:41449544ae9f64f6a6e6e7ae69f24c0fd9049e38beef4bba2bbed3e3e730e98c",
+        "matrixRow": "2026-07-26-kungfu-linux-product-qualification"
+      },
+      "claimBoundary": "The predecessor seal closes the prior Linux Assignment; each later terminal cut must still provide its own exact-head platform evidence."
     },
     {
       "id": "windows-native-and-product",
-      "state": "retained-dependency",
-      "evidence": [],
+      "state": "proved",
+      "evidence": [
+        "framework/maintainability/terminal-evidence-matrix.json"
+      ],
       "dependentAssignment": {
         "assignmentId": "2026-07-26-kungfu-windows-product-qualification",
         "requestRoot": "sha256:2abb9fb07ac8c09ca1658388731b23a1fd47bbe8115ec284cc24c54e93b2157c",
         "captureReceiptRoot": "sha256:bc649a491246506f2e0067b089af5539136ebce8972270c2783650ca8f7b1b04"
       },
-      "claimBoundary": "Static shifu.cmd parity is not Windows execution evidence; exact Windows native and product execution remains a separately owned paused Assignment."
+      "closureBinding": {
+        "sealedStateRoot": "sha256:41449544ae9f64f6a6e6e7ae69f24c0fd9049e38beef4bba2bbed3e3e730e98c",
+        "matrixRow": "2026-07-26-kungfu-windows-product-qualification"
+      },
+      "claimBoundary": "The predecessor seal closes the prior Windows Assignment; static parity never substitutes for exact-head Windows execution on a later cut."
     },
     {
       "id": "physical-durability",
@@ -540,6 +564,50 @@
         "scripts/readonly-agent-bootstrap.test.mjs"
       ],
       "rollback": "Restore the projection function in the full verifier and remove the read-only invariant route."
+    },
+    {
+      "id": "tui-control-plane-state",
+      "original": [
+        "framework/tui/src/profile-shell.tsx"
+      ],
+      "boundary": "Input transitions, quick-action matching, and focus fencing are pure state authority; Ink rendering remains in the profile shell.",
+      "target": "framework/tui/src/control-plane-state.ts",
+      "owner": "framework/tui",
+      "dependencyInvariant": "The state module imports no Ink or React runtime and the profile shell preserves its public exports.",
+      "tests": [
+        "framework/tui/src/control-plane.test.ts",
+        "framework/tui/src/profile-shell.test.ts"
+      ],
+      "rollback": "Restore the pure state declarations in profile-shell.tsx and remove the compatibility re-export."
+    },
+    {
+      "id": "affected-native-artifact-lookup",
+      "original": [
+        "scripts/affected-native-proof.mjs"
+      ],
+      "boundary": "GitHub artifact discovery and producer selection are release-evidence reads; proof construction and sealing remain in the affected-native authority.",
+      "target": "framework/release/affected-native-artifact-lookup.mjs",
+      "owner": "framework/release",
+      "dependencyInvariant": "Discovery admits only exact repository identities, trusted workflow events, successful runs, bounded age, and exact merge-group heads.",
+      "tests": [
+        "scripts/affected-native-proof.test.mjs"
+      ],
+      "rollback": "Move the remote lookup functions back without changing their public affected-native exports."
+    },
+    {
+      "id": "merge-queue-delivery-evidence",
+      "original": [
+        "scripts/cancel-dequeued-merge-group-runs.mjs"
+      ],
+      "boundary": "Dequeue settlement mutates queue cleanup state; retained delivery evidence reconstruction is a read-only maintainability projection.",
+      "target": "framework/maintainability/delivery-evidence.mjs",
+      "owner": "framework/maintainability",
+      "dependencyInvariant": "The evidence reader consumes validated affected-native attempts and GitHub observations without owning dequeue mutation.",
+      "tests": [
+        "scripts/candidate-timeline-events.test.mjs",
+        "scripts/cancel-dequeued-merge-group-runs.test.mjs"
+      ],
+      "rollback": "Move the read-only evidence helpers back and preserve the existing re-export surface."
     }
   ]
 }
