+++ framework/maintainability/terminal-evidence-matrix.json
@@ -1,11 +1,14 @@
 {
-  "schema": "kungfu.maintainability-terminal-evidence-matrix/v1",
-  "goalId": "2026-07-27-kungfu-maintainability-terminal-closure",
+  "schema": "kungfu.maintainability-terminal-evidence-matrix/v2",
+  "goalId": "2026-07-29-kungfu-terminal-review-remediation",
+  "predecessor": {
+    "assignmentId": "2026-07-27-kungfu-maintainability-terminal-closure",
+    "sealedStateRoot": "sha256:41449544ae9f64f6a6e6e7ae69f24c0fd9049e38beef4bba2bbed3e3e730e98c"
+  },
   "sourceBinding": {
     "repository": "kungfu-systems/kungfu",
     "protectedBranch": "dev/v4/v4.0",
-    "deliveryPullRequest": 1578,
-    "exactHeadRule": "Evaluate this matrix only from the exact protected commit being qualified. Every workflow selector, artifact coordinate, completion claim, and review must resolve to that commit or to a listed already-merged prerequisite.",
+    "exactHeadRule": "The verifier reads Git HEAD, the protected ref, and a retained GitHub delivery snapshot. Evidence never chooses its own source, pull request, or merge identity.",
     "alreadyMergedPrerequisites": [
       {
         "pullRequest": 1553,
@@ -18,6 +21,8 @@
     ]
   },
   "terminalEvidence": {
+    "schema": "kungfu.maintainability-terminal-evidence/v2",
+    "verifier": "framework/maintainability/terminal-evidence.mjs",
     "sourceWorkflow": ".github/workflows/affected-native-pr.yml",
     "platformPreflightWorkflow": ".github/workflows/alpha-promotion-preflight.yml",
     "productWorkflow": ".github/workflows/build.yml",
@@ -35,7 +40,7 @@
       "macos-arm64",
       "windows-x64"
     ],
-    "artifactIdentityRule": "Use the Build workflow artifact coordinates and digests whose publish-source-sha equals the exact protected commit.",
+    "artifactIdentityRule": "Every platform run and artifact resolves to the live protected commit, and every declared sha256 root recomputes from retained bytes.",
     "reviewRule": {
       "reviewer": "kungfu-origin",
       "decision": "APPROVED",
@@ -43,13 +48,18 @@
       "freshContext": true,
       "maximumOpenSeverity": "no P0, no P1, and no material P2 in scope"
     },
-    "assignmentRule": "Every listed child and the parent must have an accepted native completion decision whose gitCommit is the exact protected commit, except the Dogfood prerequisite may additionally bind its listed protected merge commits and installed-product root."
+    "assignmentRule": "The current Assignment request, accepted completion decision, and sealed state are retained and exact-head bound. Predecessor rows are closed projections of the listed immutable predecessor seal, not actionable work."
   },
   "exceptions": [],
   "rows": [
     {
       "assignmentId": "2026-07-26-kungfu-authority-convergence",
-      "disposition": "implemented",
+      "disposition": "closed",
+      "nativeClosure": {
+        "requestRoot": "sha256:3c21a18843d601a08012bce640fa6c184358179cc9f82999d7420b80d7806b97",
+        "captureReceiptRoot": "sha256:e9668f6ac8cf97bb59db87c1c5ec04076dd14b4d294abc1a25408c2d2dfb552b",
+        "sealedBy": "predecessor.sealedStateRoot"
+      },
       "obligations": [
         "one declared authority per changed semantic family",
         "retire duplicated KFX forwarding decisions",
@@ -75,7 +85,12 @@
     },
     {
       "assignmentId": "2026-07-26-kungfu-responsibility-hotspot-decomposition",
-      "disposition": "implemented",
+      "disposition": "closed",
+      "nativeClosure": {
+        "requestRoot": "sha256:5f71d58aac57e46f268dd1953473249aab182e0e3f8061a7d75929c3f81be9ee",
+        "captureReceiptRoot": "sha256:9694edded6e090e588c56286ebf9c957718ecfb14b3e92b8b74e7ffbcace2bde",
+        "sealedBy": "predecessor.sealedStateRoot"
+      },
       "obligations": [
         "extract cohesive complexity-governance ownership",
         "extract cohesive Buildchain KFD runtime ownership",
@@ -83,7 +98,10 @@
       ],
       "sourceEvidence": [
         "framework/maintainability/complexity-governance.mjs",
+        "framework/maintainability/delivery-evidence.mjs",
+        "framework/release/affected-native-artifact-lookup.mjs",
         "framework/release/buildchain-kfd-runtime.mjs",
+        "framework/tui/src/control-plane-state.ts",
         "framework/maintainability/code-complexity-policy.json"
       ],
       "rawChecks": [
@@ -100,7 +118,12 @@
     },
     {
       "assignmentId": "2026-07-26-kungfu-linux-product-qualification",
-      "disposition": "implemented",
+      "disposition": "closed",
+      "nativeClosure": {
+        "requestRoot": "sha256:abfb20e87a7c06437431c133b4e32c5cb72a8c1ca4c7e5ddb8917b7eb77e0a4f",
+        "captureReceiptRoot": "sha256:91305cbc2255ed4ef18e2f7966be6f9acbc7ea68544b1535c261a70cebfa8974",
+        "sealedBy": "predecessor.sealedStateRoot"
+      },
       "obligations": [
         "bind native, CLI, API, package, and product evidence to one exact Linux artifact",
         "prove locked-checkout zero-write discovery",
@@ -123,7 +146,12 @@
     },
     {
       "assignmentId": "2026-07-26-kungfu-windows-product-qualification",
-      "disposition": "implemented",
+      "disposition": "closed",
+      "nativeClosure": {
+        "requestRoot": "sha256:2abb9fb07ac8c09ca1658388731b23a1fd47bbe8115ec284cc24c54e93b2157c",
+        "captureReceiptRoot": "sha256:bc649a491246506f2e0067b089af5539136ebce8972270c2783650ca8f7b1b04",
+        "sealedBy": "predecessor.sealedStateRoot"
+      },
       "obligations": [
         "execute Windows native and product qualification rather than infer it from shell parity",
         "prove platform-appropriate locked-checkout zero-write behavior",
@@ -146,7 +174,7 @@
     },
     {
       "assignmentId": "2026-07-26-kungfu-dogfood-operational-hardening",
-      "disposition": "implemented",
+      "disposition": "closed",
       "obligations": [
         "diagnose exact Profile-root drift without mutation",
         "separate authorized recovery from read-only diagnosis",
@@ -171,7 +199,7 @@
     },
     {
       "assignmentId": "2026-07-27-kungfu-complexity-gate-integrity",
-      "disposition": "implemented",
+      "disposition": "closed",
       "obligations": [
         "recompute measurements at exact base and candidate commits",
         "bind independent approvals, complete path sets, dates, expiry, and removal conditions",
@@ -199,7 +227,7 @@
     },
     {
       "assignmentId": "2026-07-27-kungfu-readonly-source-acceptance-closure",
-      "disposition": "implemented",
+      "disposition": "closed",
       "obligations": [
         "classify every supported source and Agent read route",
         "run the full cold source suite without checkout writes, network access, or dependency installation",
